+++ src/main/java/com/direwolf20/buildinggadgets/common/network/PacketChangeRange.java
@@ -53,9 +53,6 @@ public class PacketChangeRange implements IMessage {
             } else if (heldItem.getItem() instanceof GadgetExchanger) {
                 GadgetExchanger gadgetExchanger = (GadgetExchanger) (heldItem.getItem());
                 gadgetExchanger.rangeChange(playerEntity, heldItem);
-            } else if (heldItem.getItem() instanceof GadgetCopyPaste) {
-                GadgetCopyPaste gadgetCopyPaste = (GadgetCopyPaste) (heldItem.getItem());
-                gadgetCopyPaste.rotateOrMirrorBlocks(heldItem, playerEntity);
             } else if (heldItem.getItem() instanceof GadgetDestruction) {
                 GadgetDestruction gadgetDestruction = (GadgetDestruction) (heldItem.getItem());
                 gadgetDestruction.switchOverlay(playerEntity, heldItem);