+++ src/main/java/com/direwolf20/buildinggadgets/common/tools/GadgetUtils.java
@@ -33,6 +33,9 @@ import net.minecraft.server.MinecraftServer;
 import net.minecraft.tileentity.TileEntity;
 import net.minecraft.util.EnumActionResult;
 import net.minecraft.util.EnumFacing;
+import net.minecraft.util.EnumFacing.Axis;
+import net.minecraft.util.Mirror;
+import net.minecraft.util.Rotation;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.util.math.RayTraceResult;
 import net.minecraft.util.text.TextComponentString;
@@ -204,6 +207,18 @@ public class GadgetUtils {
             return 1;
         }
         return tagCompound.getInteger("range");
+    }
+
+    public static IBlockState rotateOrMirrorBlock(EntityPlayer player, IBlockState state) {
+        if (player.isSneaking())
+            return state.withMirror(player.getHorizontalFacing().getAxis() == Axis.X ? Mirror.LEFT_RIGHT : Mirror.FRONT_BACK);
+
+        return state.withRotation(Rotation.CLOCKWISE_90);
+    }
+
+    public static void rotateOrMirrorToolBlock(ItemStack stack, EntityPlayer player) {
+        setToolBlock(stack, rotateOrMirrorBlock(player, getToolBlock(stack)));
+        setToolActualBlock(stack, rotateOrMirrorBlock(player, getToolActualBlock(stack)));
     }
 
     private static void setToolBlock(ItemStack stack, @Nullable IBlockState state) {